+++ examples/ThirdPartyLibs/Gwen/Controls/VerticalScrollBar.cpp
@@ -40,8 +40,8 @@ void VerticalScrollBar::Layout(Skin::Base* skin)
 	//Calculate bar sizes
 	float barHeight = (m_fViewableContentSize / m_fContentSize) * (float)(Height() - (GetButtonSize() * 2));
 
-	if (barHeight < GetButtonSize() * 0.5f)
-		barHeight = GetButtonSize() * 0.5f;
+	if (barHeight < (float)GetButtonSize() * 0.5f)
+		barHeight = (float)GetButtonSize() * 0.5f;
 
 	m_Bar->SetHeight((int)barHeight);
 	m_Bar->SetHidden((float)(Height() - (GetButtonSize() * 2)) <= barHeight);
